+++ src/Controller/BridgeBelote/DefaultController.php
@@ -1,51 +1,22 @@
 <?php namespace App\Controller\BridgeBelote;
 
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use App\Controller\Application\GameController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Twig\Environment;
-use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
 
-use Vankosoft\ApplicationBundle\Component\Context\ApplicationContextInterface;
-
-class DefaultController extends AbstractController
+class DefaultController extends GameController
 {
-    /** @var ApplicationContextInterface */
-    private $applicationContext;
-    
-    /** @var Environment */
-    private $templatingEngine;
-    
-    /** @var EntityRepository */
-    private $gamesRepository;
-    
-    public function __construct(
-        ApplicationContextInterface $applicationContext,
-        Environment $templatingEngine,
-        EntityRepository $gamesRepository
-    ) {
-        $this->applicationContext   = $applicationContext;
-        $this->templatingEngine     = $templatingEngine;
-        $this->gamesRepository      = $gamesRepository;
-    }
-    
     public function index( Request $request ): Response
     {
         $gameSlug   = 'bridge-belote';
-        $game   = $this->gamesRepository->findOneBy( ['slug' => $gameSlug] );
+        $game       = $this->gamesRepository->findOneBy( ['slug' => $gameSlug] );
+        $signature  = $this->getVerifySignature();
         
-        return new Response( $this->templatingEngine->render( $this->getTemplate(), ['game' => $game] ) );
-    }
-    
-    protected function getTemplate(): string
-    {
-        $template   = 'bridge-belote/Pages/BridgeBelote/index.html.twig';
-        
-        $appSettings    = $this->applicationContext->getApplication()->getSettings();
-        if ( ! $appSettings->isEmpty() && $appSettings[0]->getTheme() ) {
-            $template   = 'Pages/BridgeBelote/index.html.twig';
-        }
-        
-        return $template;
+        return new Response(
+            $this->templatingEngine->render( $this->getTemplate( $gameSlug , 'Pages/BridgeBelote/index.html.twig' ), [
+                'game'                  => $game,
+                'urlLoginBySignature'   => $signature ? $signature->getSignedUrl() : null,
+            ])
+        );
     }
 }