+++ src/Entity/UserManagement/User.php
@@ -3,12 +3,40 @@
 use Doctrine\ORM\Mapping as ORM;
 use Vankosoft\UsersBundle\Model\User as BaseUser;
 
-/**
- * @ORM\Entity
- * @ORM\Table(name="VSUM_Users")
- */
-class User extends BaseUser
+use Vankosoft\UsersSubscriptionsBundle\Model\Interfaces\SubscribedUserInterface;
+use Vankosoft\UsersSubscriptionsBundle\Model\Traits\SubscribedUserEntity;
+use Vankosoft\PaymentBundle\Model\Interfaces\UserPaymentAwareInterface;
+use Vankosoft\PaymentBundle\Model\Traits\UserPaymentAwareEntity;
+use Vankosoft\PaymentBundle\Model\Interfaces\CustomerInterface;
+use Vankosoft\PaymentBundle\Model\Traits\CustomerEntity;
+use Vankosoft\CatalogBundle\Model\Interfaces\UserSubscriptionAwareInterface;
+use Vankosoft\CatalogBundle\Model\Traits\UserSubscriptionAwareEntity;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+
+#[ORM\Entity]
+#[ORM\Table(name: "VSUM_Users")]
+class User extends BaseUser implements
+    SubscribedUserInterface,
+    UserPaymentAwareInterface,
+    CustomerInterface,
+    UserSubscriptionAwareInterface
 {
+    use SubscribedUserEntity;
+    use UserPaymentAwareEntity;
+    use CustomerEntity;
+    use UserSubscriptionAwareEntity;
+    
+    public function __construct()
+    {
+        $this->newsletterSubscriptions  = new ArrayCollection();
+        $this->orders                   = new ArrayCollection();
+        $this->pricingPlanSubscriptions = new ArrayCollection();
+        
+        parent::__construct();
+    }
+    
     /**
      * {@inheritDoc}
      */