+++ themes/GamePlatform_AngularJs/assets/js/games/application/constants.ts
@@ -1,0 +1,8 @@
+/**
+ * https://medium.com/codex/how-to-share-constants-in-typescript-project-8f76a2e40352
+ */
+export class AppConstants
+{
+    static readonly RESPONSE_STATUS_OK: string      = 'ok';
+    static readonly RESPONSE_STATUS_ERROR: string   = 'error';
+};