+++ themes/GamePlatform_AngularJs/assets/js/games/application/interfaces/auth.ts
@@ -1,6 +1,7 @@
 export interface IAuth
 {
     id: number;
+    
     email: string;
     username: string;