+++ themes/GamePlatform_AngularJs/assets/js/games/bridge-belote/bridge-belote.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit, OnDestroy, Inject } from '@angular/core';
+import { Component, OnInit, OnDestroy, Inject, ElementRef, isDevMode } from '@angular/core';
 
 import { AuthService } from '../application/services/auth.service'
+import { ApiService } from '../application/services/api.service'
 
 import cssCardGameString from '../application/assets/CardGame.scss'
 import cssGameString from './bridge-belote.component.scss'
@@ -20,16 +21,32 @@ declare var $: any;
 })
 export class BridgeBeloteComponent implements OnInit, OnDestroy
 {
-    isLoggedIn: boolean = false;
+    urlLoginBySignature?: string;
+    
+    isLoggedIn: boolean         = false;
+    developementClass: string   = '';
     
     constructor(
-        @Inject(AuthService) private authStore: AuthService
+        @Inject(ElementRef) private elementRef: ElementRef,
+        @Inject(AuthService) private authStore: AuthService,
+        @Inject(ApiService) private apiService: ApiService
     ) {
-        /* */
+        if( isDevMode() ) {
+            this.developementClass  = 'developement';
+        }
+        
+        this.urlLoginBySignature = this.elementRef.nativeElement.getAttribute( 'urlLoginBySignature' );
+        
         this.authStore.isLoggedIn().subscribe( ( isLoggedIn: boolean ) => {
+            //alert( isLoggedIn );
             this.isLoggedIn = isLoggedIn;
         });
     
+        if ( ! this.isLoggedIn && this.urlLoginBySignature?.length ) {
+            this.apiService.loginBySignedUrl( this.urlLoginBySignature )
+        }
+        
+        //this.debugApplication();
     }
     
     ngOnInit(): void
@@ -41,4 +58,15 @@ export class BridgeBeloteComponent implements OnInit, OnDestroy
     {
 
     }
+    
+    debugApplication()
+    {
+        if ( this.urlLoginBySignature?.length ) {
+            alert( this.urlLoginBySignature );
+        } else {
+            alert( 'Missing Login By Signature URL !!!' );
+        }
+        
+        alert( this.isLoggedIn );
+    }  
 }